Move contact related objects required for client to client project
