+++ WeddingWebsite/Client/Client.csproj
@@ -21,8 +21,4 @@
       <_ContentIncludedByDefault Remove="wwwroot\index.html" />
     </ItemGroup>
 
-    <ItemGroup>
-        <Folder Include="Models\" />
-    </ItemGroup>
-
 </Project>
